Refactor done
